Removed Location class.

+++ lib/types/constants.dart
@@ -1,9 +1,9 @@
 class Constants {
   // Game.
-  static const int gridStart = -3;
-  static const int gridEnd = 3;
-  static const int gridSize = gridEnd - gridStart + 1;
-  static const int cellOffset = 2;
+  static const double gridStart = -3.0;
+  static const double gridEnd = 3.0;
+  static const double gridSize = gridEnd - gridStart + 1;
+  static const double cellOffset = 2.0;
    
   // BoxWidgets.
   static const double boxPadding = 4.0;
